+ span for composite source + separate implementation for net5-6
