+++ test/DotNetty.Handlers.Tests/TlsHandlerTest.cs
@@ -69,131 +69,6 @@ namespace DotNetty.Handlers.Tests
                 select new object[] { frameLengths, isClient, writeStrategyFactory(), protocol.Item1, protocol.Item2 };
         }
 
-        [Fact]
-        public async Task TlsRead2()
-        {
-            var executor = new SingleThreadEventExecutor("test executor", TimeSpan.FromMilliseconds(10));
-
-            try
-            {
-                var writeTasks = new List<Task>();
-                Tuple<EmbeddedChannel, SslStream> setup = await SetupStreamAndChannelAsync(
-                    isClient: true,
-                    executor,
-                    new AsIsWriteStrategy(),
-                    serverProtocol: SslProtocols.Tls12,
-                    clientProtocol: SslProtocols.Tls12,
-                    writeTasks: writeTasks);
-            }
-            finally
-            {
-                while (TlsHandler.Events.TryDequeue(out string msg))
-                {
-                    this.Output.WriteLine(msg);
-                }
-                await executor.ShutdownGracefullyAsync(TimeSpan.Zero, TimeSpan.Zero);
-            }
-        }
-
-        [Fact]
-        public async Task TlsRead1()
-        {
-            while (TlsHandler.Events.TryDequeue(out _))
-            {
-            }
-            
-            var executor = new SingleThreadEventExecutor("test executor", TimeSpan.FromMilliseconds(10));
-
-            try
-            {
-                X509Certificate2 tlsCertificate = TestResourceHelper.GetTestCertificate();
-                string targetHost = tlsCertificate.GetNameInfo(X509NameType.DnsName, false);
-
-                var tlsHandler = new TlsHandler(
-                    stream => new SslStream(stream, true, (sender, certificate, chain, errors) => true),
-                    new ClientTlsSettings(SslProtocols.Tls11, false, new List<X509Certificate>(), targetHost)
-                );
-
-                TlsHandler.Trace("Test" + nameof(this.TlsRead1), "Register handler");
-                var ch = new EmbeddedChannel(tlsHandler);
-                var writeStrategy = new AsIsWriteStrategy();
-
-                IByteBuffer readResultBuffer = Unpooled.Buffer(4 * 1024);
-
-                var writeTasks = new List<Task>();
-                var mediationStream = new MediationStream(
-                    async output =>
-                    {
-                        TlsHandler.Trace("Test" + nameof(this.TlsRead1), "TestMediationStream.ReadFunc");
-                        if (writeTasks.Count > 0)
-                        {
-                            await Task.WhenAll(writeTasks).WithTimeout(TimeSpan.FromSeconds(2));
-                            writeTasks.Clear();
-                        }
-
-                        if (readResultBuffer.ReadableBytes < output.Count)
-                        {
-                            if (ch.Active)
-                            {
-                                var buffer = ch.ReadOutbound<IByteBuffer>();
-                                if (buffer != null)
-                                {
-                                    if (buffer.IsReadable())
-                                    {
-                                        readResultBuffer.WriteBytes(buffer);
-                                    }
-
-                                    buffer.Release();
-                                }
-                            }
-                        }
-
-                        int read = Math.Min(output.Count, readResultBuffer.ReadableBytes);
-                        readResultBuffer.ReadBytes(output.Array, output.Offset, read);
-                        return read;
-                    },
-                    input =>
-                    {
-                        TlsHandler.Trace("Test" + nameof(this.TlsRead1), "TestMediationStream.WriteFunc");
-                        Task task = executor.SubmitAsync(
-                            () =>
-                            {
-                                TlsHandler.Trace("Test" + nameof(this.TlsRead1), "TestMediationStream.WriteFunc in executor thread");
-                                return writeStrategy.WriteToChannelAsync(ch, input);
-                            }).Unwrap();
-                        writeTasks.Add(task);
-                        return task;
-                    },
-                    () => ch.CloseAsync()
-                );
-
-                var serverSsl = new SslStream(mediationStream, true, (_1, _2, _3, _4) => true);
-                TlsHandler.Trace("Test" + nameof(this.TlsRead1), "AuthenticateAsServerAsync");
-                await Task.Run(() => serverSsl.AuthenticateAsServerAsync(tlsCertificate, false, SslProtocols.Tls11, false).WithTimeout(TimeSpan.FromSeconds(2)));
-                writeTasks.Clear();
-
-                await tlsHandler.HandshakeCompletedFuture.WithTimeout(TimeSpan.FromSeconds(2));
-            }
-            catch (Exception)
-            {
-                while (TlsHandler.Events.TryDequeue(out string msg))
-                {
-                    this.Output.WriteLine(msg);
-                }
-
-                throw;
-            }
-            finally
-            {
-                // while (TlsHandler.Events.TryDequeue(out string msg))
-                // {
-                //     this.Output.WriteLine(msg);
-                // }
-                await executor.ShutdownGracefullyAsync(TimeSpan.Zero, TimeSpan.Zero);
-            }
-        }
-
-
         [Theory]
         [MemberData(nameof(GetTlsReadTestData))]
         public async Task TlsRead(int[] frameLengths, bool isClient, IWriteStrategy writeStrategy, SslProtocols serverProtocol, SslProtocols clientProtocol)
@@ -202,7 +77,6 @@ namespace DotNetty.Handlers.Tests
             {
             }
 
-            
             this.Output.WriteLine($"frameLengths: {string.Join(", ", frameLengths)}");
             this.Output.WriteLine($"isClient: {isClient}");
             this.Output.WriteLine($"writeStrategy: {writeStrategy}");
